requirements.txt
requests==2.31.0 
psycopg2-binary==2.9.11
python-dotenv==1.0.0 
numpy
pandas
